requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-community
pydantic
duckdb
chromadb
python-dotenv
yfinance
pandas
pandas_ta
scikit-learn
statsforecast
xgboost
tavily-python
streamlit
plotly
